Relationship between turnout and right-wing-populist-parties
This project contains data from 7 EU-Member-States (AT, CH, DE, DK, FR,
NL and UK) and the results from latest elections (two national elections
between 2009-2016 by country). The scripts "datenaufbereitung.R" in each
folder creates a dataset of small-scale data (mostly constituencies or
political districts) and the change in turnout and in the result of a
populist-right-wing-party from the earlier election to the later one.

The analyse.R-Scripts generates Scatterplots via ggplot2.
The gfx_europe.R-script generates a grid-graph from all of these gfx.

The "wanderungen"-Folder contains the data for voter-shifting between
parties in all federal-state-elections in Germany in 2016. The script
generates a bar-graph showing the shifts form every political group to
the AfD.

The data was used in two articles which you can find under the following
links.